new updates

+++ src/main/java/com/example/RecipeApplication/repository/RecipesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 
-public interface RecipesRepository extends CrudRepository<Recipe, Long> {
+public interface RecipesRepository extends CrudRepository<Recipe, Integer> {
 }
